inventory scripts function was moved to entity module;
